+++ src/gui/ObjectiveTracker/mixins/ProfessionsTrackerModule.lua
@@ -1,6 +1,6 @@
 ---@class Addon
 local Addon = select(2, ...)
-local GUI, Recipes, Util = Addon.GUI, Addon.Recipes, Addon.Util
+local GUI, Orders, Recipes, Util = Addon.GUI, Addon.Orders, Addon.Recipes, Addon.Util
 
 ---@class GUI.ObjectiveTracker.ProfessionsTrackerModule
 ---@field module ObjectiveTrackerModuleMixin
@@ -69,7 +69,14 @@ end
 ---------------------------------------
 
 function Self:OnAddonLoaded()
-    Recipes:RegisterCallback(Recipes.Event.TrackedUpdated, self.module.MarkDirty, self.module)
-    Recipes:RegisterCallback(Recipes.Event.TrackedAmountUpdated, self.module.MarkDirty, self.module)
-    Recipes:RegisterCallback(Recipes.Event.TrackedQualityUpdated, self.module.MarkDirty, self.module)
+    local MarkDirty = Util:FnBind(self.module.MarkDirty, self.module)
+
+    Recipes:RegisterCallback(Recipes.Event.TrackedUpdated, MarkDirty)
+    Recipes:RegisterCallback(Recipes.Event.TrackedAmountUpdated, MarkDirty)
+    Recipes:RegisterCallback(Recipes.Event.TrackedQualityUpdated, MarkDirty)
+    Recipes:RegisterCallback(Recipes.Event.TrackedAllocationUpdated, MarkDirty)
+
+    Orders:RegisterCallback(Orders.Event.TrackedUpdated, MarkDirty)
+    Orders:RegisterCallback(Orders.Event.TrackedAmountUpdated, MarkDirty)
+    Orders:RegisterCallback(Orders.Event.TrackedAllocationUpdated, MarkDirty)
 end