+++ src/modules/Orders.lua
@@ -1,15 +1,19 @@
 ---@class Addon
 local Addon = select(2, ...)
-local Recipes, Util = Addon.Recipes, Addon.Util
+local Recipes, Reagents, Util = Addon.Recipes, Addon.Reagents, Addon.Util
 
 ---@class Orders: CallbackRegistryMixin
 ---@field Event Orders.Event
 local Self = Mixin(Addon.Orders, CallbackRegistryMixin)
 
----@type table<boolean, CraftingOrderInfo[]>
+---@type table<boolean, CraftingOrderInfo[][]>
 Self.tracked = { [false] = {}, [true] = {} }
----@type boolean[]
-Self.creatingProvidedSlots = {}
+---@type number[]
+Self.trackedAmounts = {}
+---@type table<boolean, RecipeAllocation[]>
+Self.trackedAllocations = { [false] = {}, [true] = {} }
+---@type true[][]
+Self.creatingProvided = {}
 
 ---------------------------------------
 --              Tracking
@@ -17,42 +21,115 @@ Self.creatingProvidedSlots = {}
 
 -- Get
 
----@param order CraftingOrderInfo
-function Self:IsTracked(order)
-    local trackedOrder = self:GetTracked(order)
-    if not trackedOrder or not Recipes:IsTracked(order) then return false end
-    return trackedOrder.orderID == order.orderID
+---@param recipeOrOrder RecipeOrOrder
+---@param isRecraft? boolean
+function Self:IsTracked(recipeOrOrder, isRecraft)
+    local orderID = type(recipeOrOrder) == "table" and recipeOrOrder.orderID or 0
+    local trackedOrders = self:GetTracked(recipeOrOrder, isRecraft)
+
+    return trackedOrders and trackedOrders[orderID] ~= nil or false
 end
 
----@param recipeOrOrder CraftingRecipeSchematic | TradeSkillRecipeInfo | CraftingOrderInfo | number
-function Self:GetTracked(recipeOrOrder)
-    local recipeID, isRecraft = Recipes:GetRecipeInfo(recipeOrOrder)
+---@param recipeOrOrder RecipeOrOrder
+---@param isRecraft? boolean
+function Self:GetTracked(recipeOrOrder, isRecraft)
+    local recipeID, isRecraft = Recipes:GetRecipeInfo(recipeOrOrder, isRecraft)
+
     return self.tracked[isRecraft][recipeID]
 end
 
-function Self:GetTrackedProvidedReagentAmounts()
-    ---@type number[]
-    local reagents = {}
-    for order in self:Enumerate() do
-        if self:IsCreating(order) then
-            local amount = Recipes:GetTrackedAmount(order) or 1
-            if amount > 0 then
-                local recipe = C_TradeSkillUI.GetRecipeSchematic(order.spellID, order.isRecraft)
 
-                for slotIndex in pairs(self.creatingProvidedSlots) do
-                    local reagent = recipe.reagentSlotSchematics[slotIndex]
-                    local itemID, quantity = reagent.reagents[1].itemID, reagent.quantityRequired ---@cast itemID -?
-                    reagents[itemID] = (reagents[itemID] or 0) + quantity * amount
-                end
-            end
-        else
+---@param recipeOrOrder RecipeOrOrder
+---@param isRecraft? boolean
+function Self:GetTrackedAmount(recipeOrOrder, isRecraft)
+    local recipeID, isRecraft = Recipes:GetRecipeInfo(recipeOrOrder, isRecraft)
+
+    local trackedOrders = self:GetTracked(recipeOrOrder, isRecraft)
+    if not trackedOrders then return end
+
+    local amount = Util:TblCount(trackedOrders)
+
+    if isRecraft then return amount end
+
+    return amount - 1 + (Self.trackedAmounts[recipeID] or 1)
+end
+
+---@param recipeOrOrder RecipeOrOrder
+---@param isRecraft? boolean
+function Self:GetTrackedAllocation(recipeOrOrder, isRecraft)
+    local recipeID, isRecraft = Recipes:GetRecipeInfo(recipeOrOrder, isRecraft)
+
+    if not self.trackedAllocations[isRecraft][recipeID] then return end
+
+    return self.trackedAllocations[isRecraft][recipeID][recipeOrOrder.orderID or 0]
+end
+
+function Self:GetTrackedReagentAmounts()
+    ---@type number[]
+    local reagents, provided = {}, {}
+
+    for order in self:Enumerate() do repeat
+        local amount = self:IsCreating(order) and self:GetTrackedAmount(order) or 1
+        if amount <= 0 then break end
+
+        local recipe = C_TradeSkillUI.GetRecipeSchematic(order.spellID, order.isRecraft)
+        local allocation = self:GetTrackedAllocation(order)
+
+        -- Provided by customer
+        if not self:IsCreating(order) then
             for _,reagent in pairs(order.reagents) do
                 local itemID, quantity = reagent.reagent.itemID, reagent.reagent.quantity
-                reagents[itemID] = (reagents[itemID] or 0) + quantity
+                provided[itemID] = (provided[itemID] or 0) + quantity
             end
         end
-    end
-    return reagents
+
+        for slotIndex,reagent in pairs(recipe.reagentSlotSchematics) do
+            local required = reagent.required and reagent.quantityRequired or 0
+            local missing = amount * required
+
+            if self:IsCreatingProvided(order, slotIndex) then
+                -- Provided by crafter
+                local itemID = reagent.reagents[1].itemID ---@cast itemID -?
+                provided[itemID] = (provided[itemID] or 0) + missing
+            elseif not Reagents:IsProvidedByOrder(reagent, order) then
+                -- Allocated
+                if allocation and allocation[slotIndex] then
+                    for _, alloc in allocation[slotIndex]:Enumerate() do
+                        missing = missing - amount * alloc.quantity
+
+                        local itemID = alloc.reagent.itemID ---@cast itemID -?
+                        reagents[itemID] = (reagents[itemID] or 0) + amount * alloc.quantity
+                    end
+                end
+
+                if missing > 0 then
+                    local itemID = reagent.reagents[1].itemID ---@cast itemID -?
+                    reagents[itemID] = (reagents[itemID] or 0) + missing
+                end
+            end
+        end
+    until true end
+
+    return reagents, provided
+end
+
+function Self:GetTrackedResultAmounts()
+    ---@type number[]
+    local items = {}
+
+    for order in self:Enumerate() do repeat
+        if not self:IsCreating(order) then break end
+
+        local amount = self:GetTrackedAmount(order)
+        if amount <= 0 then break end
+
+        local output = C_TradeSkillUI.GetRecipeOutputItemData(order.spellID, nil, nil, order.minQuality)
+        if not output or not output.itemID then break end
+
+        items[output.itemID] = (items[output.itemID] or 0) + amount
+    until true end
+
+    return items
 end
 
 -- Set
@@ -62,17 +139,63 @@ end
 function Self:SetTracked(order, value)
     value = value ~= false
 
-    if value and self:IsTracked(order) then return end
-    if not value and not self:GetTracked(order) then return end
+    if value == self:IsTracked(order) then return end
 
     local recipeID, isRecraft = Recipes:GetRecipeInfo(order)
-    self.tracked[isRecraft][recipeID] = value and order or nil
+    local orderID = order.orderID or 0
 
-    if value then Recipes:SetTracked(order, true) end
+    if value and not self.tracked[isRecraft][recipeID] then self.tracked[isRecraft][recipeID] = {} end
+
+    local list = self.tracked[isRecraft][recipeID]
+    list[orderID] = value and order or nil
+
+    if not value then
+        if not next(list) then self.tracked[isRecraft][recipeID] = nil end
+        self:SetTrackedAllocation(order, nil)
+        if self:IsCreating(order) then self:SetTrackedAmount(order, nil) end
+    elseif not Recipes:IsTracked(order) then
+        Recipes:SetTracked(order, true)
+        Recipes:SetTrackedAmount(order, 0)
+    end
 
     self:TriggerEvent(self.Event.TrackedUpdated, order, value)
 
     self:UpdateCreatingReagents()
+end
+
+---@param recipeOrOrder RecipeOrOrder
+---@param amount? number
+function Self:SetTrackedAmount(recipeOrOrder, amount)
+    if amount and amount < 0 then amount = 0 end
+    if amount == 1 then amount = nil end
+
+    local recipeID = Recipes:GetRecipeInfo(recipeOrOrder)
+    if self.trackedAmounts[recipeID] == amount then return end
+
+    self.trackedAmounts[recipeID] = amount
+
+    self:TriggerEvent(Self.Event.TrackedAmountUpdated, recipeID, amount)
+end
+
+---@param recipeOrOrder RecipeOrOrder
+---@param allocation? RecipeAllocation
+---@param isRecraft? boolean
+function Self:SetTrackedAllocation(recipeOrOrder, allocation, isRecraft)
+    local recipeID, isRecraft = Recipes:GetRecipeInfo(recipeOrOrder, isRecraft)
+    local orderID = type(recipeOrOrder) == "table" and recipeOrOrder.orderID or 0
+
+    if not self.trackedAllocations[isRecraft][recipeID] then
+        if not allocation then return end
+        self.trackedAllocations[isRecraft][recipeID] = {}
+    end
+
+    self.trackedAllocations[isRecraft][recipeID][orderID] = allocation
+
+    if not allocation and not next(self.trackedAllocations[isRecraft][recipeID]) then
+        self.trackedAllocations[isRecraft][recipeID] = nil
+    end
+
+    self:TriggerEvent(Self.Event.TrackedAllocationUpdated, recipeID, isRecraft, orderID, allocation)
 end
 
 -- Clear
@@ -100,18 +223,18 @@ function Self:IsCreating(order)
     return order and not order.orderID
 end
 
-function Self:GetCreating()
-    for _,orders in pairs(self.tracked) do
-        local order = Util:TblWhere(orders, "orderID", nil)
-        if order then return order end
-    end
+---@param slot ReagentSlot
+function Self:IsCreatingSlotProvided(slot)
+    return slot.Checkbox:IsShown() and not slot.Checkbox:GetChecked()
 end
 
 ---@param order? CraftingOrderInfo
 ---@param slotIndex number
 function Self:IsCreatingProvided(order, slotIndex)
     if not self:IsCreating(order) then return false end ---@cast order -?
-    if Self.creatingProvidedSlots[slotIndex] then return true end
+
+    local creatingProvided = Self.creatingProvided[order.spellID]
+    if creatingProvided then return creatingProvided[slotIndex] or false end
 
     local frame = ProfessionsCustomerOrdersFrame
     if not frame or frame.Form.order ~= order then return false end
@@ -121,7 +244,7 @@ function Self:IsCreatingProvided(order, slotIndex)
 
     for slot in frame.Form.reagentSlotPool:EnumerateActive() do
         if slot:GetSlotIndex() == slotIndex then
-            return slot.Checkbox:IsShown() and not slot.Checkbox:GetChecked()
+            return self:IsCreatingSlotProvided(slot)
         end
     end
 
@@ -131,26 +254,31 @@ end
 ---@param slot ReagentSlot
 ---@param silent? boolean
 function Self:UpdateCreatingReagent(slot, silent)
-    local reagent = slot:GetReagentSlotSchematic()
-    local provided = slot.Checkbox:IsShown() and not slot.Checkbox:GetChecked()
+    local tx, reagent = slot:GetTransaction(), slot:GetReagentSlotSchematic()
+    local recipe = tx:GetRecipeSchematic()
+    local provided = self:IsCreatingSlotProvided(slot)
 
-    self.creatingProvidedSlots[reagent.slotIndex] = provided and reagent.required or nil
+    if not Self.creatingProvided[recipe.recipeID] then
+        if not provided then return end
+        Self.creatingProvided[recipe.recipeID] = {}
+    end
+
+    Self.creatingProvided[recipe.recipeID][reagent.slotIndex] = provided or nil
+
+    if not next(Self.creatingProvided[recipe.recipeID]) then Self.creatingProvided[recipe.recipeID] = nil end
 
     if silent then return end
 
-    self:TriggerEvent(self.Event.CreatingReagentsUpdated)
+    self:TriggerEvent(self.Event.CreatingReagentsUpdated, reagent.slotIndex)
 end
 
 function Self:UpdateCreatingReagents()
     if not ProfessionsCustomerOrdersFrame then return end
 
-    wipe(self.creatingProvidedSlots)
-
     local form = ProfessionsCustomerOrdersFrame.Form
-    local recipe = form.transaction:GetRecipeSchematic()
-    local tracked = recipe and self:GetTracked(form.order) == form.order
+    local tracked = self:IsTracked(form.order)
 
-    if recipe and tracked then
+    if tracked then
         for slot in form.reagentSlotPool:EnumerateActive() do
             self:UpdateCreatingReagent(slot, true)
         end
@@ -165,7 +293,7 @@ end
 
 ---@return fun(): CraftingOrderInfo?
 function Self:Enumerate()
-    return Util:TblEnum(self.tracked, 2)
+    return Util:TblEnum(self.tracked, 3)
 end
 
 ---------------------------------------
@@ -174,9 +302,11 @@ end
 
 ---@class Orders.Event
 ---@field TrackedUpdated "TrackedUpdated"
+---@field TrackedAmountUpdated "TrackedAmountUpdated"
+---@field TrackedAllocationUpdated "TrackedAllocationUpdated"
 ---@field CreatingReagentsUpdated "CreatingReagentsUpdated"
 
-Self:GenerateCallbackEvents({ "TrackedUpdated", "CreatingReagentsUpdated" })
+Self:GenerateCallbackEvents({ "TrackedUpdated", "TrackedAmountUpdated", "TrackedAllocationUpdated", "CreatingReagentsUpdated" })
 Self:OnLoad()
 
 ---@param recipeID number