+++ CHANGES.md
@@ -1,1 +1,3 @@
+- Add tracking multiple orders of the same recipe
 - Properly handle finishing reagents that change resourcefulness/multicraft yield
+- Updated reagent, perk and recipe data